Improve bootstrap to always be safe

+++ bootstrap.sh
@@ -12,25 +12,32 @@ symlink() {
     if [[ ! -d $(dirname $new) ]]; then
       echo mkdir -p $(dirname $new)
     fi
-    echo "rm -f $new"
+    [[ -f $new ]] && echo "rm -f $new"
     echo "ln -sf $existing $new"
   else
     mkdir -p $(dirname $new)
-    rm -f $new
+    [[ -f $new ]] && rm -f $new
     ln -sf $existing $new
   fi
 }
 
+dirty() {
+  [[ $(git diff --shortstat 2> /dev/null | tail -n1) != ""  ]] && return 0
+}
+
 update() {
-  git pull origin master && \
-    git submodule init && \
-    git submodule update
+  if ! dirty; then
+    git pull origin master && \
+      git submodule init && \
+      git submodule update
+  fi
 }
 
 install() {
   for f in $(find . -type f -not -wholename '*.git/*' -not -name 'bootstrap.sh' -not -name 'README.md'); do
     symlink $f $dest/${f#./}
   done
+  update
   echo "Done!"
 }
 
@@ -39,10 +46,11 @@ if [[ $1 == "--dry" ]]; then
   set -- "$2"
 fi
 
+
 case $1 in
   install) install ;;
   update) update ;;
-  *) echo "Usage: bootstrap.sh [install|update]"; exit 1
+  *) echo "Usage: bootstrap.sh [install|update]"; exit 1 ;;
 esac
 
 popd > /dev/null
